+++ app/controllers/exam_templates_controller.rb
@@ -80,8 +80,8 @@ class ExamTemplatesController < ApplicationController
   end
 
   def generate
-    copies = params[:numCopies].to_i
-    index = params[:examTemplateIndex].to_i
+    copies = params[:exam_template][:num_copies].to_i
+    index = params[:exam_template][:start_index].to_i
     assignment = Assignment.find(params[:assignment_id])
     exam_template = assignment.exam_templates.find(params[:id])
 
@@ -118,22 +118,18 @@ class ExamTemplatesController < ApplicationController
   def add_fields
     assignment = Assignment.find(params[:assignment_id])
     exam_template = assignment.exam_templates.find(params[:id])
-    if params[:automatic_parsing] == 'true'
-      exam_template.automatic_parsing = true
-      exam_template.cover_fields = params[:cover_fields]
-      exam_template.crop_x = params[:x].to_f
-      exam_template.crop_y = params[:y].to_f
-      exam_template.crop_width = params[:width].to_f
-      exam_template.crop_height = params[:height].to_f
+    if params[:exam_template][:automatic_parsing] == '1'
+      exam_template.update(exam_template_crop_fields_params)
     else
-      exam_template.automatic_parsing = false
-      exam_template.cover_fields = ''
-      exam_template.crop_x = nil
-      exam_template.crop_y = nil
-      exam_template.crop_width = nil
-      exam_template.crop_height = nil
+      exam_template.update(
+        automatic_parsing: false,
+        cover_fields: '',
+        crop_x: nil,
+        crop_y: nil,
+        crop_width: nil,
+        crop_height: nil
+      )
     end
-    exam_template.save
     redirect_to action: 'index'
   end
 
@@ -304,5 +300,17 @@ class ExamTemplatesController < ApplicationController
          :name,
          template_divisions_attributes: [:id, :start, :end, :label, :_destroy]
        )
+  end
+
+  def exam_template_crop_fields_params
+    params.require(:exam_template)
+          .permit(
+            :automatic_parsing,
+            :cover_fields,
+            :crop_x,
+            :crop_y,
+            :crop_width,
+            :crop_height
+          )
   end
 end